+++ requirements-test.txt
@@ -3,3 +3,4 @@ pytest-flask==1.2.0
 pytest-cov==4.1.0
 pytest-env==1.0.1
 pytest==7.4.3
+pre-commit==3.5.0
