+++ run_worker.sh
@@ -2,7 +2,7 @@
 
 ./yaptide/admin/simulators.py install --name shieldhit --path /simulators/shieldhit12a/bin
 
-if [ -z "$S3_TOPAS_BUCKET" ] || [ -z "$S3_TOPAS_KEY" ] || [ -z "$S3_TOPAS_VERSION" ] || [ -z "$S3_GEANT_BUCKET" ]; then
+if [ -z "$S3_TOPAS_BUCKET" ] || [ -z "$S3_TOPAS_KEY" ] || [ -z "$S3_TOPAS_VERSION" ] || [ -z "$S3_GEANT4_BUCKET" ]; then
         echo "One or more environment variables required by TOPAS are not set, skipping TOPAS installation"
     else
         apt-get update
