+++ tests/test_download_shieldhit.py
@@ -4,31 +4,31 @@ import subprocess
 import sys
 import pytest
 from pathlib import Path
-from yaptide.admin.simulators import download_shieldhit_from_s3
+from yaptide.admin.simulator_storage import download_shieldhit_from_s3
 
 
 def check_if_environment_variables_set() -> bool:
     """Check if environment variables are set"""
     result = True
-    if not 'S3_ENDPOINT' in os.environ:
-        result = False
-    if not 'S3_ACCESS_KEY' in os.environ:
-        result = False
-    if not 'S3_SECRET_KEY' in os.environ:
-        result = False
-    if not 'S3_SHIELDHIT_BUCKET' in os.environ:
-        result = False
-    if not 'S3_SHIELDHIT_KEY' in os.environ:
-        result = False
+    for var_name in ['S3_ENDPOINT', 'S3_ACCESS_KEY', 'S3_SECRET_KEY', 'S3_SHIELDHIT_BUCKET', 'S3_SHIELDHIT_KEY']:
+        if var_name not in os.environ:
+            logging.error(' variable %s not set', var_name)
+            result = False
     return result
+
 
 @pytest.mark.skipif(sys.platform == "win32", reason="Lets not tests this on Windows.")
 def test_if_shieldhit_downloaded(tmpdir, shieldhit_binary_filename):
     """Check if SHIELD-HIT12A binary is downloaded and can be executed"""
     if check_if_environment_variables_set():
-        bucket = os.getenv("S3_SHIELDHIT_BUCKET")
-        key = os.getenv("S3_SHIELDHIT_KEY")
-        assert download_shieldhit_from_s3(bucket=bucket, key=key, shieldhit_path=tmpdir) is True
+        assert download_shieldhit_from_s3(endpoint=os.getenv("S3_ENDPOINT"),
+                                          access_key=os.getenv("S3_ACCESS_KEY"),
+                                          secret_key=os.getenv("S3_SECRET_KEY"),
+                                          bucket=os.getenv("S3_SHIELDHIT_BUCKET"),
+                                          key=os.getenv("S3_SHIELDHIT_KEY"),
+                                          password=os.getenv("S3_ENCRYPTION_PASSWORD"),
+                                          salt=os.getenv("S3_ENCRYPTION_SALT"),
+                                          shieldhit_path=tmpdir) is True
         expected_path = Path(tmpdir / shieldhit_binary_filename)
         assert expected_path.exists(), "Expected path does not exist."
         assert expected_path.stat().st_size > 0, "Expected path is empty."
@@ -38,16 +38,16 @@ def test_if_shieldhit_downloaded(tmpdir, shieldhit_binary_filename):
             # If check=True and the exit code is non-zero, raises a
             # CalledProcessError (has return code and output/error streams).
             # text=True means stdout and stderr will be strings instead of bytes
-            completed_process = subprocess.run(command, 
-                                               check=True, 
-                                               stdout=subprocess.PIPE, 
-                                               stderr=subprocess.PIPE, 
+            completed_process = subprocess.run(command,
+                                               check=True,
+                                               stdout=subprocess.PIPE,
+                                               stderr=subprocess.PIPE,
                                                text=True)
-            
+
             # Capture stdout and stderr
             command_stdout = completed_process.stdout
             command_stderr = completed_process.stderr
-            
+
             # Log stdout and stderr using logging
             logging.info("Command Output:\n%s", command_stdout)
             logging.info("Command Error Output:\n%s", command_stderr)
